Split out functions and constants to lib.rs.
